rewritten Tests for FESolvePoisson (based on already available PROBLEMdefinitions), deactivated some other tests which have to be rewritten too

+++ src/PROBLEMdefinitions/POISSON_1D_polynomials.jl
@@ -1,4 +1,4 @@
-function getProblemData(polynomial_coefficients::Vector{Int64})
+function getProblemData(polynomial_coefficients::Vector{Float64})
 
     # auto-computing coefficients of - 2nd derivative
     polynomial_coefficients_rhs = zeros(Float64,length(polynomial_coefficients)-2)
